+++ debug_toolbar/panels/request.py
@@ -59,13 +59,12 @@ class RequestPanel(Panel):
         self.record_stats(view_info)
 
         if hasattr(request, "session"):
-            self.record_stats(
-                {
-                    "session": {
-                        "list": [
-                            (k, request.session.get(k))
-                            for k in sorted(request.session.keys())
-                        ]
-                    }
-                }
-            )
+            try:
+                session_list = [
+                    (k, request.session.get(k)) for k in sorted(request.session.keys())
+                ]
+            except TypeError:
+                session_list = [
+                    (k, request.session.get(k)) for k in request.session.keys()
+                ]
+            self.record_stats({"session": {"list": session_list}})